Add workflow to create Python binaries and build Refstudio for supported OS (#508)
Adds a GH action (reuse cut-release.yml) to build refstudio for both mac architectures, and release (as draft) to github.
The release version is read from the package.json. For now we added the next version (23.2.0) with an -alpha.1 suffix.

This PR also fixes an issue with the react-pdf in intel chips. We've had to add polyfill and use the legacy pdf worker.
Additionally add some extra startup logs (main window -> splashscreen) to debug.


---------

Co-authored-by: Carlos Guedes <cguedes@gmail.com>